+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support Token base security authentication
 - Support for Scheduled Autoscaling
+- Support for `dbVersion` field in Autonomous databases Container database resources
 
 ## 3.80.0 (June 10, 2020)